desktop/requirements.txt
customtkinter>=5.2.0
cryptography>=42.0.0
Pillow>=10.0.0
qrcode[pil]>=7.4.0
mysql-connector-python==8.3.0

# API-сервер (для мобильного клиента)
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
